refactor: replace base doc url
We add this to catch any usage of the documentation URL we've missed while reverse engineering the code.

To ensure that the user is supported by IONOS in case of questions about the product or unexpected behavior,
we replace all links to the Nextcloud Docs that the user may encounter with links to the IONOS Help Center.

If a legacy theme is enabled, Nextcloud uses the theme's methods to generate links to documentation

Signed-off-by: Kai Henseler <kai.henseler@strato.de>