fix(template): Use `<div>` instead of `<main>` to support Vue3 apps mounting Vue3 does not replace the element while mounting but only renders within (replace inner HTML). So it would result in two stacked `<main>` elements which is invalid and an accessibility issue. Instead we just use a `<div>` element for mounting.
For Vue2 apps this does not change anything as the whole element will be
replaced with a new `<main>` element.
For vanilla apps this will slightly decrease the accessibility as the
main landmark is now missing, but this is not a hard accessibility issue
as it would be for Vue3 apps having two main elements.

(cherry picked from commit https://github.com/nextcloud/server/commit/182acb1e299f71623609bde7ebca21f71985befd)
Signed-off-by: Misha M.-Kupriyanov <kupriyanov@strato.de>